begin to organize fields in geometries
