+++ inc/geometries/init.h
@@ -5,6 +5,8 @@
  *
  * Initialize and Damping objects
  */
+namespace dg
+{
 namespace solovev
 {
 ///@addtogroup profiles
@@ -428,83 +430,8 @@ struct ZonalFlow
     Psi psip_;
 };
 
-/**
- * @brief testfunction to test the parallel derivative 
-      \f[ f(R,Z,\varphi) = -\frac{\cos(\varphi)}{R\hat b_\varphi} \f]
- */ 
-template<class Collective>
-struct TestFunction
-{
-    TestFunction( const Collective& c, double R0) :  
-        bhatR_(c, R0),
-        bhatZ_(c, R0),
-        bhatP_(c, R0) {}
-    /**
-     * @brief \f[ f(R,Z,\varphi) = -\frac{\cos(\varphi)}{R\hat b_\varphi} \f]
-     */ 
-    double operator()( double R, double Z, double phi)
-    {
-//         return psip_(R,Z,phi)*sin(phi);
-//         double Rmin = gp_.R_0-(p_.boxscaleRm)*gp_.a;
-//         double Rmax = gp_.R_0+(p_.boxscaleRp)*gp_.a;
-//         double kR = 1.*M_PI/(Rmax - Rmin);
-//         double Zmin = -(p_.boxscaleZm)*gp_.a*gp_.elongation;
-//         double Zmax = (p_.boxscaleZp)*gp_.a*gp_.elongation;
-//         double kZ = 1.*M_PI/(Zmax - Zmin);
-        double kP = 1.;
-//         return sin(phi*kP)*sin((R-Rmin)*kR)*sin((Z-Zmin)*kZ); //DIR
-//         return cos(phi)*cos((R-Rmin)*kR)*cos((Z-Zmin)*kZ);
-//         return sin(phi*kP); //DIR
-//         return cos(phi*kP); //NEU
-        return -cos(phi*kP)/bhatP_(R,Z,phi)/R; //NEU 2
-
-    }
-    private:
-    BHatR<Collective> bhatR_;
-    BHatZ<Collective> bhatZ_;
-    BHatP<Collective> bhatP_;
-};
-/**
- * @brief analyitcal solution of the parallel derivative of the testfunction
- *  \f[ \nabla_\parallel(R,Z,\varphi) f = \frac{\sin(\varphi)}{R}\f]
- */ 
-template<class Collective>
-struct DeriTestFunction
-{
-    DeriTestFunction( const Collective& c, double R0) :
-        bhatR_(c, R0),
-        bhatZ_(c, R0),
-        bhatP_(c, R0) {}
-/**
- * @brief \f[ \nabla_\parallel f = \frac{\sin(\varphi)}{R}\f]
- */ 
-    double operator()( double R, double Z, double phi)
-    {
-//         double Rmin = gp_.R_0-(p_.boxscaleRm)*gp_.a;
-//         double Rmax = gp_.R_0+(p_.boxscaleRp)*gp_.a;
-//         double kR = 1.*M_PI/(Rmax - Rmin);
-//         double Zmin = -(p_.boxscaleZm)*gp_.a*gp_.elongation;
-//         double Zmax = (p_.boxscaleZp)*gp_.a*gp_.elongation;
-//         double kZ = 1.*M_PI/(Zmax - Zmin);
-        double kP = 1.;
-//          return (bhatR_(R,Z,phi)*sin(phi)*sin((Z-Zmin)*kZ)*cos((R-Rmin)*kR)*kR+
-//                 bhatZ_(R,Z,phi)*sin(phi)*sin((R-Rmin)*kR)*cos((Z-Zmin)*kZ)*kZ+
-//                 bhatP_(R,Z,phi)*cos(phi)*sin((R-Rmin)*kR)*sin((Z-Zmin)*kZ)*kP); //DIR
-//         return -bhatR_(R,Z,phi)*cos(phi)*cos((Z-Zmin)*kZ)*sin((R-Rmin)*kR)*kR-
-//                bhatZ_(R,Z,phi)*cos(phi)*cos((R-Rmin)*kR)*sin((Z-Zmin)*kZ)*kZ-
-//                bhatP_(R,Z,phi)*sin(phi)*cos((R-Rmin)*kR)*cos((Z-Zmin)*kZ)*kP;
-//         return  bhatP_(R,Z,phi)*cos(phi*kP)*kP; //DIR
-//         return  -bhatP_(R,Z,phi)*sin(phi*kP)*kP; //NEU
-        return sin(phi*kP)*kP/R; //NEU 2
-
-    }
-    private:
-    BHatR<Collective> bhatR_;
-    BHatZ<Collective> bhatZ_;
-    BHatP<Collective> bhatP_;
-};
-
 
 ///@}
 }//namespace solovev
+}//namespace dg
